chore(refactor): move commands into bin
remove one layer of DI
