+++ bin/cmds/common-args.js
